refactor(tests,benches,docs): converge the estate on one shape
primary-store.md rewritten against the verified journal_covers reality
(memory/mapped/forked/recording all journal; only disk whole-graph
checkpoints — the old three-case list was wrong on two); data-loading
ladder 4->3 rungs; structural-sharing.md gains the cell-journal successor
section. Tracked bench cells 27->24: regime-duplicate cells deleted,
save cells honestly renamed (values preserved), the changed-operation
cell removed from both baselines rather than carrying an incomparable
anchor; mapped bench cell no longer measures fsync (durable='off' with
reason). Source quality green by tightening: rollback::apply,
unpack_column, write_disk_snapshot factored, two exception entries
REMOVED; rollback_tests split into a module dir. test_columnar_storage
-> test_column_store_integrity, test_incremental_columnar ->
test_incremental_columns. Track E mapped/disk ownership arms added,
red-first (the mapped arm pins allocation identity — the stronger
invariant). CHANGELOG reordered as one release story, entries
byte-identical.
