+++ CYPHER.md
@@ -962,7 +962,7 @@ int / float / bool / string. Invalid JSON or a non-string argument returns
 a JSON string rather than as graph structure.
 
 The code graph keeps `Function.parameters`, `Class.fields`, and
-`Function.signature` as JSON (the columnar store holds scalars only), so
+`Function.signature` as JSON (property columns hold scalars only), so
 `parse_json` is how you query inside them:
 
 ```python
