+++ crates/kglite/src/graph/dir_graph/disk_persistence.rs
@@ -167,42 +167,7 @@ impl DirGraph {
     }
 
     fn write_disk_snapshot(&mut self, dir: &std::path::Path) -> Result<(), String> {
-        // Build CSR from pending edges if not yet built.
-        self.ensure_disk_edges_built()?;
-        // Merge overflow edges back so conn_type_index and
-        // peer_count_histogram reflect every live edge. Skipped during
-        // builds; done here as a one-shot so users only pay the cost at
-        // save time, not per add_connections batch.
-        //
-        // The seal path consumes `overflow_out` / `overflow_in` directly,
-        // while a rewrite must compact first so every derived index is rebuilt
-        // from the complete CSR. Use the lower layer's target-aware decision:
-        // generation stages are distinct directories and always rewrite.
-        if let GraphBackend::Disk(ref mut dg) = self.graph {
-            dg.begin_persist();
-            let disposition = dg.save_disposition(dir);
-            if disposition == crate::graph::storage::disk::graph_persist::SaveDisposition::Rewrite
-                && dg.has_overflow()
-            {
-                dg.compact()
-                    .map_err(|e| format!("disk compaction failed: {e}"))?;
-            }
-            // Auto-build the cross-type global title index so that
-            // `MATCH (n {title: 'X'})` and `g.search(text)` are O(log N)
-            // out of the box on every saved disk graph. Runs after
-            // CSR / overflow consolidation so it sees the final node
-            // set. Tied to `save_disk` rather than `build_csr_*` so
-            // node-only graphs (no edges) still get the index built.
-            dg.build_global_property_index("title")
-                .map_err(|e| format!("title index build failed: {e}"))?;
-            // Likewise index `nid` — the string id form for prefixed-id
-            // datasets (Wikidata `"Q42"`). Since 0.11.0 `{nid: 'Q42'}` is a
-            // plain string-property lookup (not the integer id-index), so the
-            // index keeps it O(log N) instead of a 124M-row scan. No-op when
-            // no type has a `nid` column.
-            dg.build_global_property_index("nid")
-                .map_err(|e| format!("nid index build failed: {e}"))?;
-        }
+        self.consolidate_disk_for_save(dir)?;
 
         // save_to_dir needs &mut access so the edge-property store can
         // drop its base mmap before overwriting (PR2).
@@ -251,36 +216,102 @@ impl DirGraph {
         // older JSON representation remains a read-only data fallback.
         crate::graph::io::file::write_interner_bin(dir, self)?;
 
-        // Save column stores (per type, sidecar format). Two modes:
+        self.write_unified_column_file(dir)?;
+        self.write_column_sidecars(dir)?;
+
+        // 0.8.13: type_indices uses a flat CSR binary keyed by interner
+        // hashes. 0.8.28+: id_indices uses an mmap-resident raw `.bin`
+        // layout — load reads via memory-mapped binary search, no eager
+        // HashMap rebuild. The loader can still read the earlier flat-CSR
+        // sidecars when the mmap files are absent; pre-0.14 bincode caches
+        // are ignored and rebuilt.
+        crate::graph::storage::disk::type_index::write_type_indices_bin(
+            dir,
+            &self.type_indices,
+            &self.interner,
+        )?;
+        crate::graph::storage::disk::id_index::write_id_indices_bin(
+            dir,
+            &self.id_indices,
+            &self.interner,
+        )?;
+
+        // Save embeddings if any (matches write_kgl behavior for in-memory saves).
+        // BTreeMap view for byte-determinism — same rationale as write_kgl.
+        if !self.embeddings.is_empty() {
+            let ordered: std::collections::BTreeMap<_, _> = self.embeddings.iter().collect();
+            write_compressed_disk_serde(dir, "embeddings.bin.zst", &ordered, "embeddings")?;
+        }
+
+        // Save timeseries_store if any (BTreeMap view for byte-determinism).
+        if !self.timeseries_store.is_empty() {
+            let ordered: std::collections::BTreeMap<_, _> = self.timeseries_store.iter().collect();
+            write_compressed_disk_serde(dir, "timeseries.bin.zst", &ordered, "timeseries")?;
+        }
+
+        // Root metadata is the graph-level completion marker. Publish it only
+        // after every required sidecar has been written successfully.
+        std::fs::write(dir.join("metadata.json"), meta_json)
+            .map_err(|e| format!("Failed to publish metadata: {}", e))?;
+
+        Ok(())
+    }
+
+    /// Bring a disk graph's CSR, overflow edges and global indexes to their
+    /// final saved state before anything is written.
+    fn consolidate_disk_for_save(&mut self, dir: &std::path::Path) -> Result<(), String> {
+        // Build CSR from pending edges if not yet built.
+        self.ensure_disk_edges_built()?;
+        // Merge overflow edges back so conn_type_index and
+        // peer_count_histogram reflect every live edge. Skipped during
+        // builds; done here as a one-shot so users only pay the cost at
+        // save time, not per add_connections batch.
         //
-        // 1. **No `columns.bin`** — DirGraph → disk saves that never
-        //    went through the N-Triples streaming builder. Write every
-        //    column store as a sidecar.
-        //
-        // 2. **`columns.bin` exists** — N-Triples-built disk graphs.
-        //    The single-file v3 layout covers every type that was
-        //    present at ingest time, and reloading it via the mmap fast
-        //    path (`file.rs:580`) is dramatically cheaper than walking
-        //    per-type sidecars on a 88k-type wiki graph. BUT: types
-        //    added post-build via `add_nodes` / `add_node` are not in
-        //    `columns.bin` and were silently dropped on save before
-        //    this fix. Emit sidecars *only* for those types — keeps the
-        //    fast path for initial types, makes mutation persistence
-        //    correct for new ones.
-        //
-        // 0.8.12 phase-1: PR1 phase 4 moved `columns.bin` under
-        // `seg_000/`, so the presence check covers both the root (legacy
-        // flat layout) and `seg_000/` (post-phase-4 segmented layout).
-        // Mode-3 (new in 0.9.15): no preexisting `columns.bin` AND
-        // we have in-memory `column_stores` (typical fresh save:
-        // streaming carve, save_subset, mutation persist of an
-        // in-memory build). Emit the unified mega-file format that
-        // the loader's mmap fast path consumes — same layout the
-        // ntriples builder produces — so the saved graph loads with
-        // the same speed as a freshly-built one. Without this, a
-        // saved DiskGraph fell into the per-type zstd sidecar path
-        // and took ~70 s to load on a 17 M-node Wikidata carve vs.
-        // ~150 ms for the full graph.
+        // The seal path consumes `overflow_out` / `overflow_in` directly,
+        // while a rewrite must compact first so every derived index is rebuilt
+        // from the complete CSR. Use the lower layer's target-aware decision:
+        // generation stages are distinct directories and always rewrite.
+        if let GraphBackend::Disk(ref mut dg) = self.graph {
+            dg.begin_persist();
+            let disposition = dg.save_disposition(dir);
+            if disposition == crate::graph::storage::disk::graph_persist::SaveDisposition::Rewrite
+                && dg.has_overflow()
+            {
+                dg.compact()
+                    .map_err(|e| format!("disk compaction failed: {e}"))?;
+            }
+            // Auto-build the cross-type global title index so that
+            // `MATCH (n {title: 'X'})` and `g.search(text)` are O(log N)
+            // out of the box on every saved disk graph. Runs after
+            // CSR / overflow consolidation so it sees the final node
+            // set. Tied to `save_disk` rather than `build_csr_*` so
+            // node-only graphs (no edges) still get the index built.
+            dg.build_global_property_index("title")
+                .map_err(|e| format!("title index build failed: {e}"))?;
+            // Likewise index `nid` — the string id form for prefixed-id
+            // datasets (Wikidata `"Q42"`). Since 0.11.0 `{nid: 'Q42'}` is a
+            // plain string-property lookup (not the integer id-index), so the
+            // index keeps it O(log N) instead of a 124M-row scan. No-op when
+            // no type has a `nid` column.
+            dg.build_global_property_index("nid")
+                .map_err(|e| format!("nid index build failed: {e}"))?;
+        }
+        Ok(())
+    }
+
+    /// Write the unified `columns.bin` mega-file for a graph that has none.
+    ///
+    /// Mode 3 (0.9.15): a fresh save — streaming carve, `save_subset`, or the
+    /// mutation persist of an in-memory build — emits the same layout the
+    /// ntriples builder produces, so the saved graph reloads through the mmap
+    /// fast path. Without it a saved `DiskGraph` fell back to per-type zstd
+    /// sidecars and took ~70 s to load on a 17 M-node Wikidata carve against
+    /// ~150 ms for the full graph.
+    ///
+    /// A pre-existing `columns.bin` (root or `seg_000/`, the pre- and
+    /// post-phase-4 layouts) means the ntriples builder already wrote one;
+    /// [`Self::write_column_sidecars`] covers the types added since.
+    fn write_unified_column_file(&self, dir: &std::path::Path) -> Result<(), String> {
         let preexisting_columns_bin =
             dir.join("seg_000/columns.bin").exists() || dir.join("columns.bin").exists();
         if !preexisting_columns_bin && self.column_store_count() > 0 {
@@ -292,46 +323,55 @@ impl DirGraph {
             crate::graph::io::unified_columns::write_unified_columns(dir, &stores, &self.interner)
                 .map_err(|e| format!("unified columns write failed: {}", e))?;
         }
+        Ok(())
+    }
 
-        let columns_meta_path = {
-            let seg0_bin = dir.join("seg_000/columns_meta.bin.zst");
-            let seg0_json = dir.join("seg_000/columns_meta.json");
-            let root_bin = dir.join("columns_meta.bin.zst");
-            let root_json = dir.join("columns_meta.json");
-            if seg0_bin.exists() {
-                Some(seg0_bin)
-            } else if seg0_json.exists() {
-                Some(seg0_json)
-            } else if root_bin.exists() {
-                Some(root_bin)
-            } else if root_json.exists() {
-                Some(root_json)
-            } else {
-                None
-            }
+    /// Which node types the graph's `columns.bin` already covers.
+    ///
+    /// Read from whichever `columns_meta` sidecar exists — the binary form and
+    /// the older JSON one, at the segmented (`seg_000/`) or the legacy flat
+    /// root. An absent sidecar means no type is covered.
+    fn types_in_columns_bin(
+        dir: &std::path::Path,
+    ) -> Result<std::collections::HashSet<String>, String> {
+        let columns_meta_path = [
+            "seg_000/columns_meta.bin.zst",
+            "seg_000/columns_meta.json",
+            "columns_meta.bin.zst",
+            "columns_meta.json",
+        ]
+        .into_iter()
+        .map(|name| dir.join(name))
+        .find(|path| path.exists());
+        let Some(meta_path) = columns_meta_path else {
+            return Ok(std::collections::HashSet::new());
         };
-        let types_in_columns_bin: std::collections::HashSet<String> =
-            if let Some(meta_path) = &columns_meta_path {
-                use crate::graph::io::ntriples::ColumnTypeMeta;
-                let metas: Vec<ColumnTypeMeta> =
-                    if meta_path.extension().and_then(|s| s.to_str()) == Some("zst") {
-                        let compressed = std::fs::read(meta_path)
-                            .map_err(|e| format!("read {}: {}", meta_path.display(), e))?;
-                        let bytes = zstd::decode_all(compressed.as_slice())
-                            .map_err(|e| format!("decompress columns_meta: {}", e))?;
-                        crate::graph::io::file::decode_disk_serde(&bytes, bytes.capacity() as u64)
-                            .map_err(|e| format!("parse columns_meta.bin: {}", e))?
-                    } else {
-                        let json = std::fs::read_to_string(meta_path)
-                            .map_err(|e| format!("read {}: {}", meta_path.display(), e))?;
-                        serde_json::from_str(&json)
-                            .map_err(|e| format!("parse columns_meta.json: {}", e))?
-                    };
-                metas.into_iter().map(|tm| tm.type_name).collect()
-            } else {
-                std::collections::HashSet::new()
-            };
+        use crate::graph::io::ntriples::ColumnTypeMeta;
+        let metas: Vec<ColumnTypeMeta> = if meta_path.extension().and_then(|s| s.to_str())
+            == Some("zst")
+        {
+            let compressed = std::fs::read(&meta_path)
+                .map_err(|e| format!("read {}: {}", meta_path.display(), e))?;
+            let bytes = zstd::decode_all(compressed.as_slice())
+                .map_err(|e| format!("decompress columns_meta: {}", e))?;
+            crate::graph::io::file::decode_disk_serde(&bytes, bytes.capacity() as u64)
+                .map_err(|e| format!("parse columns_meta.bin: {}", e))?
+        } else {
+            let json = std::fs::read_to_string(&meta_path)
+                .map_err(|e| format!("read {}: {}", meta_path.display(), e))?;
+            serde_json::from_str(&json).map_err(|e| format!("parse columns_meta.json: {}", e))?
+        };
+        Ok(metas.into_iter().map(|tm| tm.type_name).collect())
+    }
 
+    /// Write a per-type `columns.zst` sidecar for every type the unified
+    /// `columns.bin` does not cover.
+    ///
+    /// Types added post-build via `add_nodes` / `add_node` are absent from an
+    /// ntriples-built `columns.bin` and were silently dropped on save before
+    /// this existed. Covered types keep the fast mmap path.
+    fn write_column_sidecars(&self, dir: &std::path::Path) -> Result<(), String> {
+        let types_in_columns_bin = Self::types_in_columns_bin(dir)?;
         let columns_dir = dir.join("columns");
         let mut sidecars_written = 0usize;
         for (type_name, store) in self.column_stores_by_name() {
@@ -379,42 +419,6 @@ impl DirGraph {
                 .map_err(|e| format!("Failed to write columns: {}", e))?;
             sidecars_written += 1;
         }
-
-        // 0.8.13: type_indices uses a flat CSR binary keyed by interner
-        // hashes. 0.8.28+: id_indices uses an mmap-resident raw `.bin`
-        // layout — load reads via memory-mapped binary search, no eager
-        // HashMap rebuild. The loader can still read the earlier flat-CSR
-        // sidecars when the mmap files are absent; pre-0.14 bincode caches
-        // are ignored and rebuilt.
-        crate::graph::storage::disk::type_index::write_type_indices_bin(
-            dir,
-            &self.type_indices,
-            &self.interner,
-        )?;
-        crate::graph::storage::disk::id_index::write_id_indices_bin(
-            dir,
-            &self.id_indices,
-            &self.interner,
-        )?;
-
-        // Save embeddings if any (matches write_kgl behavior for in-memory saves).
-        // BTreeMap view for byte-determinism — same rationale as write_kgl.
-        if !self.embeddings.is_empty() {
-            let ordered: std::collections::BTreeMap<_, _> = self.embeddings.iter().collect();
-            write_compressed_disk_serde(dir, "embeddings.bin.zst", &ordered, "embeddings")?;
-        }
-
-        // Save timeseries_store if any (BTreeMap view for byte-determinism).
-        if !self.timeseries_store.is_empty() {
-            let ordered: std::collections::BTreeMap<_, _> = self.timeseries_store.iter().collect();
-            write_compressed_disk_serde(dir, "timeseries.bin.zst", &ordered, "timeseries")?;
-        }
-
-        // Root metadata is the graph-level completion marker. Publish it only
-        // after every required sidecar has been written successfully.
-        std::fs::write(dir.join("metadata.json"), meta_json)
-            .map_err(|e| format!("Failed to publish metadata: {}", e))?;
-
         Ok(())
     }
 }
