+++ docs/_generated/project-facts.md
@@ -65,4 +65,4 @@ Regenerate with `python scripts/render_docs_facts.py`. CI checks this file for d
 - CPU: `Apple M4`
 - Python: `CPython 3.14.3`
 - pytest-benchmark schema/plugin version: `5.2.3`
-- Recorded benchmarks: `27`
+- Recorded benchmarks: `24`
