+++ tests/benchmarks/baselines/current.json
@@ -778,113 +778,8 @@
     },
     {
       "group": null,
-      "name": "test_bench_columnar_enable",
-      "fullname": "tests/benchmarks/test_bench_core.py::test_bench_columnar_enable",
-      "params": null,
-      "param": null,
-      "extra_info": {},
-      "options": {
-        "disable_gc": false,
-        "timer": "perf_counter",
-        "min_rounds": 100,
-        "max_time": 1.0,
-        "min_time": 5e-06,
-        "warmup": 20
-      },
-      "stats": {
-        "min": 0.00014679203741252422,
-        "max": 0.00021079112775623798,
-        "mean": 0.00017565715109553204,
-        "stddev": 9.624999476716766e-06,
-        "rounds": 6629,
-        "median": 0.0001787920482456684,
-        "iqr": 1.7594487871974707e-05,
-        "q1": 0.00016628071898594499,
-        "q3": 0.0001838752068579197,
-        "iqr_outliers": 2,
-        "stddev_outliers": 2251,
-        "outliers": "2251;2",
-        "ld15iqr": 0.00014679203741252422,
-        "hd15iqr": 0.000210499856621027,
-        "ops": 5692.90799585008,
-        "total": 1.164431254612282,
-        "iterations": 1
-      }
-    },
-    {
-      "group": null,
-      "name": "test_bench_columnar_cypher_where",
-      "fullname": "tests/benchmarks/test_bench_core.py::test_bench_columnar_cypher_where",
-      "params": null,
-      "param": null,
-      "extra_info": {},
-      "options": {
-        "disable_gc": false,
-        "timer": "perf_counter",
-        "min_rounds": 100,
-        "max_time": 1.0,
-        "min_time": 5e-06,
-        "warmup": 20
-      },
-      "stats": {
-        "min": 0.00013045803643763065,
-        "max": 0.0003604590892791748,
-        "mean": 0.00019741505021084928,
-        "stddev": 2.410638620416203e-05,
-        "rounds": 6318,
-        "median": 0.00019810441881418228,
-        "iqr": 3.1624920666217804e-05,
-        "q1": 0.00018070894293487072,
-        "q3": 0.00021233386360108852,
-        "iqr_outliers": 81,
-        "stddev_outliers": 1930,
-        "outliers": "1930;81",
-        "ld15iqr": 0.00013400008901953697,
-        "hd15iqr": 0.0002602080348879099,
-        "ops": 5065.469927100033,
-        "total": 1.2472682872321457,
-        "iterations": 1
-      }
-    },
-    {
-      "group": null,
-      "name": "test_bench_columnar_cypher_match",
-      "fullname": "tests/benchmarks/test_bench_core.py::test_bench_columnar_cypher_match",
-      "params": null,
-      "param": null,
-      "extra_info": {},
-      "options": {
-        "disable_gc": false,
-        "timer": "perf_counter",
-        "min_rounds": 100,
-        "max_time": 1.0,
-        "min_time": 5e-06,
-        "warmup": 20
-      },
-      "stats": {
-        "min": 5.5409036576747894e-06,
-        "max": 4.187505692243576e-05,
-        "mean": 6.317109279457288e-06,
-        "stddev": 2.7382538314220957e-07,
-        "rounds": 173907,
-        "median": 6.292015314102173e-06,
-        "iqr": 1.2479722499847412e-07,
-        "q1": 6.250105798244476e-06,
-        "q3": 6.3749030232429504e-06,
-        "iqr_outliers": 4393,
-        "stddev_outliers": 3482,
-        "outliers": "3482;4393",
-        "ld15iqr": 6.082933396100998e-06,
-        "hd15iqr": 6.582820788025856e-06,
-        "ops": 158300.25344850632,
-        "total": 1.0985895234625787,
-        "iterations": 1
-      }
-    },
-    {
-      "group": null,
-      "name": "test_bench_columnar_save_kgl",
-      "fullname": "tests/benchmarks/test_bench_core.py::test_bench_columnar_save_kgl",
+      "name": "test_bench_save_kgl",
+      "fullname": "tests/benchmarks/test_bench_core.py::test_bench_save_kgl",
       "params": null,
       "param": null,
       "extra_info": {},
@@ -918,8 +813,8 @@
     },
     {
       "group": null,
-      "name": "test_bench_save_v3",
-      "fullname": "tests/benchmarks/test_bench_core.py::test_bench_save_v3",
+      "name": "test_bench_save_kgl_new_file",
+      "fullname": "tests/benchmarks/test_bench_core.py::test_bench_save_kgl_new_file",
       "params": null,
       "param": null,
       "extra_info": {},
