+++ tests/benchmarks/baselines/current.linux.json
@@ -906,113 +906,8 @@
     },
     {
       "group": null,
-      "name": "test_bench_columnar_enable",
-      "fullname": "test_bench_core.py::test_bench_columnar_enable",
-      "params": null,
-      "param": null,
-      "extra_info": {},
-      "options": {
-        "disable_gc": false,
-        "timer": "perf_counter",
-        "min_rounds": 100,
-        "max_time": 1.0,
-        "min_time": 5e-06,
-        "warmup": 20
-      },
-      "stats": {
-        "min": 0.00041970500001298205,
-        "max": 0.0005728720000206522,
-        "mean": 0.00043594926444535183,
-        "stddev": 1.3325504777342257e-05,
-        "rounds": 3392,
-        "median": 0.00043444249996582585,
-        "iqr": 1.6465500010554024e-05,
-        "q1": 0.0004258415000037985,
-        "q3": 0.0004423070000143525,
-        "iqr_outliers": 76,
-        "stddev_outliers": 436,
-        "outliers": "436;76",
-        "ld15iqr": 0.00041970500001298205,
-        "hd15iqr": 0.00046715399997765417,
-        "ops": 2293.8449070955007,
-        "total": 1.4787399049986334,
-        "iterations": 1
-      }
-    },
-    {
-      "group": null,
-      "name": "test_bench_columnar_cypher_where",
-      "fullname": "test_bench_core.py::test_bench_columnar_cypher_where",
-      "params": null,
-      "param": null,
-      "extra_info": {},
-      "options": {
-        "disable_gc": false,
-        "timer": "perf_counter",
-        "min_rounds": 100,
-        "max_time": 1.0,
-        "min_time": 5e-06,
-        "warmup": 20
-      },
-      "stats": {
-        "min": 0.0003508059999717261,
-        "max": 0.0005479150000269328,
-        "mean": 0.0003840547134107804,
-        "stddev": 1.909801525512849e-05,
-        "rounds": 2774,
-        "median": 0.0003809230000229036,
-        "iqr": 2.3745000021335727e-05,
-        "q1": 0.0003703119999727278,
-        "q3": 0.0003940569999940635,
-        "iqr_outliers": 67,
-        "stddev_outliers": 814,
-        "outliers": "814;67",
-        "ld15iqr": 0.0003508059999717261,
-        "hd15iqr": 0.0004297039999983099,
-        "ops": 2603.7956704632647,
-        "total": 1.065367775001505,
-        "iterations": 1
-      }
-    },
-    {
-      "group": null,
-      "name": "test_bench_columnar_cypher_match",
-      "fullname": "test_bench_core.py::test_bench_columnar_cypher_match",
-      "params": null,
-      "param": null,
-      "extra_info": {},
-      "options": {
-        "disable_gc": false,
-        "timer": "perf_counter",
-        "min_rounds": 100,
-        "max_time": 1.0,
-        "min_time": 5e-06,
-        "warmup": 20
-      },
-      "stats": {
-        "min": 1.3304999981755827e-05,
-        "max": 0.00014317799997343172,
-        "mean": 1.4104250745255625e-05,
-        "stddev": 2.285476845595177e-06,
-        "rounds": 70438,
-        "median": 1.373599997123165e-05,
-        "iqr": 2.210000502600451e-07,
-        "q1": 1.364499996725499e-05,
-        "q3": 1.3866000017515034e-05,
-        "iqr_outliers": 6375,
-        "stddev_outliers": 1672,
-        "outliers": "1672;6375",
-        "ld15iqr": 1.3314999989688658e-05,
-        "hd15iqr": 1.4205999946170778e-05,
-        "ops": 70900.61131651244,
-        "total": 0.9934752139943157,
-        "iterations": 1
-      }
-    },
-    {
-      "group": null,
-      "name": "test_bench_columnar_save_kgl",
-      "fullname": "test_bench_core.py::test_bench_columnar_save_kgl",
+      "name": "test_bench_save_kgl",
+      "fullname": "test_bench_core.py::test_bench_save_kgl",
       "params": null,
       "param": null,
       "extra_info": {},
@@ -1046,8 +941,8 @@
     },
     {
       "group": null,
-      "name": "test_bench_save_v3",
-      "fullname": "test_bench_core.py::test_bench_save_v3",
+      "name": "test_bench_save_kgl_new_file",
+      "fullname": "test_bench_core.py::test_bench_save_kgl_new_file",
       "params": null,
       "param": null,
       "extra_info": {},
